Fixing Bug in aggregator.py module, unit_aggregation_check function
Problem:
When aggregating on a single level instead of all levels, the new unit
of items which were not aggregated were not correctly returned. For
example using aggregate method in a Database instance by passing
levels='Consumption category' the new Database has not the units of
Activity, Sector, Commodity and ... as they were not aggregated.

Solution:
The bug was corrected by fixing function unit_aggregation_check in
mario/tools/aggregation.py by taking the default unit of measures for
the items that are not aggregated.