Updating the hybrid sut exiobase parser (issue #37)
Problem:
	mario was missing a parser to read hybrid sut exiobase. The
previus version of this database was so slow to read due to using excel
format for all the data. withing the new version, csv files are used
that makes it easier to read.
Now the Hybrid version of EXIOBASE is available on Zenodo at https://zenodo.org/record/7244919#.Y6Lr0NXMIQ8
A new parse function is needed in order to take advantage of a lighter format.

Changes:

- parserid.py
In this module a new dict is added that performs as the database
metedata. the dictionary will guide a function to read the files.
in this dictionary (called hybrid_sut_exiobase_parser_id) main matrices
and extensions are detailed with the way they should be parsed with
extra information)

- tableparser.py
function hybrid_sut_exiobase_reader is added.
The function will read and reshape the files to make it compatible with
mario format.

1. at first the function identifies the extensions passed by the user and
by using the hybrid_sut_exiobase_parser_id, identifies which file and
how to read.

2. the level of information in terms of index and columns for different
   matrices will be created

3. as the extensions are present for activities and final_consumption
   categories, and they parsed separately, they need to be concated. in
case no extension is passed, an empty dataframe will be created.

4. units that are stored as level of index in original database, needs
   to be extracted

5. By using the Use and Supply matrices, Z matrix will be created and
   the necessary matrices will be calcultaed.

6. database info will be passed

- parserclass.py
the new parser function will be given to API through a function called
(hybrid_sut_exiobase). the function returns the mario database.

- test_hybrid_exiobase.py

contains the test files for the parser.

note: the test files will download the database from the zenodo
repository and save it into an ignored file in the following directory:
tests/mock/temp_files