Updating the changes for hybrid_exiobase_parser (issue #37)
- parser_id.py
updating the parser id for reading the extensions

- parserclass.py
1. updating note on the function doc
2. adding notes to the mario.Database.meta for changing the names of the
   extensions

- tableparser.py
This exiobase database has differenet kind of extensions of the same
type. For example, there are different type of wastes such as waste_use
or waste_supply for Food. Another example is the emission. To avoid
confusion for mario and the user when reading or using the extensions,
the name of extensions are changed to have more info on the exyensions.

1. For all the extensions except Emission ones, the type of extension is
added to the end of the name inside parenthesis(line1386 to line1392) . Example:

extension row= "Aquatic plants" , extension type = "resource" -> new extension row = "Aquatic plants (resource)"

2. For emission that there is an extra level of information on
   Compartmant, will be as follow (line 1377 to line 1384):

Emission name	        	Compartment   extension type
Carbon dioxide, fossil		air 		Emiss         --> new
extension row = "Carbon dioxide, fossil (air - Emiss)"

- test_hybrid_exiobase.py
the tests are completed using a new function

test_read_extensions:
the function reads all the extensions and checks the consistency of
reading and renaming them (as discussed in the prevoius section) as well
as the units of measures using the metadata file

note: the prevoius test function (test_parse_main_data) was missing some
tests on the units of measures and region info comparison that is
completed.