Get_data function sub-optimality (issue #33).
Issue:
The get_data function was supposed to be used as a back-hand. Probabily for this reason, it looks sub-optimal and may be improved. Best way to use it to extract a specific matrix (e.g. f) of the main object (world) is:

world.get_data(matrices=['f'], scenarios=['scenario 2'])['scenario 2'][0]

The function may be improved by avoiding the pass the same info twice.
Furthermore, a dedicated documentation should be added.

Solution:
the best way to solve the problem would be to have a new function with a
more user-friendy function that executes the get_data function in a
simpler and more intuitive way and leave the get_data function as a
private function that is extensively used in the backend. Probably
changing this function itself will arise backward-compatibility issues.

A new function called "query" is added to relax the abovementioned
issues.

Changes:
1. AttrData.py
query function --> with docs

2. test_attrdata.py
test_query --> function unit test

3. minor changes for mario_test, load_test function to make it
   compatible with recent changes of make the mode as a non-default
value in excel parser function.