Parse eurostat improvememt (issue #8)
The originall function has some issues with performing the quality check
on the downloaded data, so when the user was downloading faulty data,
the parser was not correctly able to capture the mistakes and inform the
user.

Moreover, eurostat database structure has changed, and the previous
structure has drastically changed.

To solve the issues, a new userostat parser has created.

Changes
-------

1. the name of the parser function has changed to parse_eurostat_sut to
   explicitly identify that the function only parses sut eurostat table

2. mario/tools/parser_id.py

a metadata has created for the parser called eurostat_id to map the
structure of the database and definition of different sets of a valid
database.

3. mario/tools/parserclass.py
the backend parser function has been added to the parser APIs with the
new function (parse_eurostat_sut instead of parse_eurostat)

4. mario/tools/tableparser.py
the main backend function is reshaped in this module

4. mario/tools/utilities.py
a util function as added to extract the metadata from an excle
spreedsheet of eurostat dataset, called "extract_metadata_from_eurostat"

-ToDo-
1. unittest
2. modify the example in the docs