Introducing mario settings (#76)
New functionalities added to mario to give the user the opp of changing
the nomenclature and the indexing in a convineint way.

This new funcitonaliteis are implemented in mario.settings.settings.py

The settings will be communicated using a yaml file, called
settings.yaml that maps the mario convensions to user preferences.

Changes in the code:

All over the mario source code, direct use of matrix names and indexings
are removed, and instead an _ENUM constant object is used, that reads
the settings. For example, if somewhere the coder needs to use X matrix,
instead of directly calling X as a str in the code, should use _ENUM.X
or _ENUM["X"].

tests

tests are added to tests/test_settings.py

docs

the API docs are added, also a comprehensive example is added to the
docs, titled "Changing mario settings".