Enhanicing the new add_sector error handling with empty files
When empty matrices are passed, pandas raises some errors. The error has
changed for the new version of pandas, and the function cannot catch
error. In 3 cases, the error can be ignored, input_from, Output_to and
final_demand.

The new changes from line 175 to 180, tries to capture this issue and
pass the erro.